rename misleading media section screen
