--- feature/home/presentation/src/main/java/com/sanaa/presentation/screen/mediaWithTypeTabScreen/MediaWithTypeTabScreenEffect.kt
@@ -1,8 +1,0 @@
-package com.sanaa.presentation.screen.mediaWithTypeTabScreen
-
-import com.sanaa.presentation.state.MediaType
-
-sealed class MediaWithTypeTabScreenEffect {
-    object NavigateBack : MediaWithTypeTabScreenEffect()
-    data class NavigateToMediaDetails(val id: Int, val mediaType: MediaType) : MediaWithTypeTabScreenEffect()
-}
